@@ -1,9 +1,15 @@
+import { Header } from './components/Header/Header'
+import Styles from './App.module.css';
+import './global.css'
+import { TasksList } from './components/TasksList/TasksList';
 
-function App() {
-
+export function App() {
   return (
-    <h1>Oi</h1>
+    <div>
+      <Header />
+      <div className={Styles.wrapper}>
+        <TasksList/>        
+      </div>
+    </div>
   )
 }
-
-export default App
